+++ tests/test_util.py
@@ -6,7 +6,6 @@ from boss.util import (
     halt,
     info,
     echo,
-    merge,
     is_iterable
 )
 
@@ -37,75 +36,6 @@ def test_info(capsys):
     out, _ = capsys.readouterr()
 
     assert message in out
-
-
-def test_merge_v0():
-    '''
-    Test for boss.util.merge()
-    Assert expected merge outcome when no conflicting keys are present.
-    '''
-    dict1 = {
-        'key1': 'value1',
-        'key2': {
-            'key3': 'value3',
-            'key4': 'value4'
-        }
-    }
-
-    dict2 = {
-        'keyA': 'valueA',
-        'key2': {
-            'keyB': 'valueB'
-        }
-    }
-
-    expectedmerge = {
-        'key1': 'value1',
-        'key2': {
-            'keyB': 'valueB',
-            'key3': 'value3',
-            'key4': 'value4'
-        },
-        'keyA': 'valueA'
-    }
-
-    merged = merge(dict1, dict2)
-
-    assert merged == expectedmerge
-
-
-def test_merge_v1():
-    '''
-    Test for boss.util.merge()
-    Assert that second dictionary overrides conflicting keys during merge
-    '''
-    dict1 = {
-        'key1': 'value1',
-        'key2': {
-            'key3': 'value3',
-            'key4': 'value4'
-        }
-    }
-
-    dict2 = {
-        'key1': 'valueA',
-        'key2': {
-            'keyB': 'valueB'
-        }
-    }
-
-    expectedmerge = {
-        'key1': 'valueA',
-        'key2': {
-            'keyB': 'valueB',
-            'key3': 'value3',
-            'key4': 'value4'
-        },
-    }
-
-    merged = merge(dict1, dict2)
-
-    assert merged == expectedmerge
 
 
 def test_is_iterable_v0():
